+++ src/pages/LoginPage.tsx
@@ -1,58 +1,71 @@
-import { signInWithEmailAndPassword } from "firebase/auth";
-import React, { useRef, useState } from "react";
-import { useNavigate } from "react-router-dom";
-import { authService } from "../Firebase";
+import React, { useState, useEffect } from "react";
+import axios from "axios";
+import { useMutation } from "react-query";
 
-const Login = () => {
-  // 메인화면 이동버튼
-  const gotoBack = () => {
-    navigate("/");
-  };
-  // 회원가입 이동버튼
-  const navigate = useNavigate();
-  const gotoSignup = () => {
-    navigate("/");
-  };
-  //아이디 인풋
-  const [userId, setUserId] = useState("");
-  const userIdRef = useRef<HTMLInputElement>(null);
-  const userIdOnChange = (event: any) => {
-    setUserId(event.target.value);
-  };
-  //패스워드 인풋
-  const [userPw, setUserPw] = useState("");
-  const userFwRef = useRef<HTMLInputElement>(null);
-  const userFwOnChange = (event: any) => {
-    setUserPw(event.target.value);
+//스팀 로그인 페이지 변수
+interface steam {
+  id: any;
+  profileimg: any;
+  nickname: any;
+}
+const STEAM_OPENID_ENDPOINT = `https://steamcommunity.com/openid/login?openid.ns=http://specs.openid.net/auth/2.0&openid.mode=checkid_setup&openid.return_to=http://localhost:3000/&openid.realm=http://localhost:3000&openid.identity=http://specs.openid.net/auth/2.0/identifier_select&openid.claimed_id=http://specs.openid.net/auth/2.0/identifier_select`;
+const LoginPage: any = () => {
+  const params: any = new URLSearchParams(window.location.search);
+  const steamId = params.get("openid.claimed_id")?.split("/")[5];
+
+  const [error, setError] = useState(null);
+  const APIKEY = "234E0113F33D5C7C4D4D5292C6774550";
+
+  const [userinfo, setUserinfo] = useState<steam>({
+    id: "",
+    profileimg: "",
+    nickname: "",
+  });
+
+  const SteamLogin = () => {
+    window.location.href = STEAM_OPENID_ENDPOINT;
   };
 
-  const loginUser = () => {
-    if (!userId) {
-      alert("아이디를 입력하세요");
-    } else if (!userPw) {
-      alert("비밀번호를 입력하세요");
-    } else {
-      alert("로그인완료");
-      signInWithEmailAndPassword(authService, userId, userPw)
-        .then(() => {
-          return;
-        })
-        .catch(() => {
-          return;
+  const userData = async () => {
+    axios
+      .get("https://api.steampowered.com/ISteamUser/GetPlayerSummaries/v2/", {
+        params: {
+          key: APIKEY,
+          //스팀로그인
+          steamids: steamId,
+        },
+      })
+      .then((response) => {
+        // console.log("userstate", response);
+        setUserinfo({
+          id: response.config.params.steamids,
+          profileimg: response?.data.response.players[0].avatarfull,
+          nickname: response?.data.response.players[0].personaname,
         });
-    }
+        console.log("data", userinfo);
+      })
+      .catch((error) => setError(error));
   };
-  console.log(authService.currentUser);
+
+  // setSteamids(testId);
+
+  useEffect(() => {
+    if (steamId) {
+      userData();
+    }
+  }, []);
+
+  const axiosInstance = axios.create({
+    baseURL: STEAM_OPENID_ENDPOINT,
+    headers: {
+      "Content-Type": "application/json",
+    },
+  });
   return (
-    <>
-      <h1>Login</h1>
-      아이디 :<input value={userId} ref={userIdRef} onChange={userIdOnChange} />
-      비밀번호 :
-      <input value={userPw} ref={userFwRef} onChange={userFwOnChange} />
-      <button onClick={loginUser}>확인</button>
-      <button onClick={gotoSignup}>회원가입</button>
-    </>
+    <div>
+      <button onClick={SteamLogin}>Login with Steam</button>
+    </div>
   );
 };
 
-export default Login;
+export default LoginPage;
