+++ src/router/Router.tsx
@@ -1,7 +1,7 @@
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 import MainPage from "../pages/MainPage";
 import SignUp from "../pages/SignUpPage";
-import ChannelSearchPage from "../pages/ChannelSearchPage";
+// import ChannelSearchPage from "../pages/ChannelSearchPage";
 import FindAccountPage from "../pages/FindAccountPage";
 import FindPasswordPage from "../pages/FindPasswordPage";
 import Login from "../pages/LoginPage";
@@ -15,11 +15,11 @@ const Router = () => {
       <GlobalStyles />
       <Routes>
         <Route path="/channelsearchpage" element={<MainPage />} />
-        <Route path="/" element={<ChannelSearchPage />} />
+        {/* <Route path="/" element={<ChannelSearchPage />} /> */}
         <Route path="/findaccountpage" element={<FindAccountPage />} />
         <Route path="/findpasswordpage" element={<FindPasswordPage />} />
         <Route path="/mypage" element={<MyPage />} />
-        <Route path="/login" element={<Login />} />
+        <Route path="/" element={<Login />} />
         <Route path="/signup" element={<SignUp />} />
         <Route path="/teamchat" element={<TeamChat />} />
       </Routes>
